break the prod/cons into modules
